Improved error handling.
Added support for appending existing CSV files.
Modified indexing of algorithms and systems for easier evaluation.

+++ de.ovgu.featureide.fm.benchmark/src/de/ovgu/featureide/fm/benchmark/process/ProcessRunner.java
@@ -1,10 +1,13 @@
 package de.ovgu.featureide.fm.benchmark.process;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
+import de.ovgu.featureide.fm.benchmark.streams.ErrStreamCollector;
+import de.ovgu.featureide.fm.benchmark.streams.ErrStreamReader;
+import de.ovgu.featureide.fm.benchmark.streams.OutStreamReader;
 import de.ovgu.featureide.fm.benchmark.streams.StreamRedirector;
-import de.ovgu.featureide.fm.benchmark.streams.StreamRedirector2;
 import de.ovgu.featureide.fm.benchmark.util.Logger;
 
 public class ProcessRunner {
@@ -12,40 +15,64 @@ public class ProcessRunner {
 	private long timeout = Long.MAX_VALUE;
 
 	public <R> void run(Algorithm<R> algorithm, Result<R> result) {
+		boolean terminatedInTime = false;
+		long startTime = 0, endTime = 0;
 		try {
 			System.gc();
 			algorithm.preProcess();
-			final List<String> command = algorithm.getCommand();
-			Logger.getInstance().logInfo(command.toString(), 1, true);
 
-			final ProcessBuilder processBuilder = new ProcessBuilder(command);
-			Process process = null;
-			long startTime = 0, endTime = 0;
-			try {
-				startTime = System.nanoTime();
-				process = processBuilder.start();
+			Logger.getInstance().logInfo(algorithm.getCommand(), 1, true);
 
-				new Thread(new StreamRedirector2(process.getInputStream(), System.out, algorithm)).start();
-				new Thread(new StreamRedirector(process.getErrorStream(), System.err)).start();
-				boolean terminatedInTime = process.waitFor(timeout, TimeUnit.MILLISECONDS);
-				endTime = System.nanoTime();
-				result.setTerminatedInTime(terminatedInTime);
-				result.setNoError(true);
-				result.setTime((endTime - startTime) / 1_000_000L);
-			} catch (Exception e) {
-				Logger.getInstance().logError(e);
-			} finally {
-				if (process != null) {
-					process.destroyForcibly();
+			final List<String> command = algorithm.getCommandElements();
+			if (!command.isEmpty()) {
+
+				final ProcessBuilder processBuilder = new ProcessBuilder(command);
+				Process process = null;
+
+				final ErrStreamCollector errStreamCollector = new ErrStreamCollector();
+				final StreamRedirector errRedirector = new StreamRedirector(
+						Arrays.asList(new ErrStreamReader(), errStreamCollector));
+				final StreamRedirector outRedirector = new StreamRedirector(
+						Arrays.asList(new OutStreamReader(), algorithm));
+				final Thread outThread = new Thread(outRedirector);
+				final Thread errThread = new Thread(errRedirector);
+				try {
+					startTime = System.nanoTime();
+					process = processBuilder.start();
+
+					outRedirector.setInputStream(process.getInputStream());
+					errRedirector.setInputStream(process.getErrorStream());
+					outThread.start();
+					errThread.start();
+
+					terminatedInTime = process.waitFor(timeout, TimeUnit.MILLISECONDS);
+					endTime = System.nanoTime();
+					result.setTerminatedInTime(terminatedInTime);
+					result.setNoError(errStreamCollector.getErrList().isEmpty());
+					result.setTime((endTime - startTime) / 1_000_000L);
+				} finally {
+					if (process != null) {
+						process.destroyForcibly();
+					}
 				}
+			} else {
+				result.setTerminatedInTime(false);
+				result.setNoError(false);
+				result.setTime(Result.INVALID_TIME);
 			}
 		} catch (Exception e) {
-			Logger.getInstance().logError(e);
+			Logger.getInstance().logError(e, true);
+			result.setTerminatedInTime(false);
+			result.setNoError(false);
+			result.setTime(Result.INVALID_TIME);
 		}
 		try {
 			result.setResult(algorithm.parseResults());
 		} catch (Exception e) {
-			Logger.getInstance().logError(e);
+			Logger.getInstance().logError(e, true);
+			if (terminatedInTime) {
+				result.setNoError(false);
+			}
 		}
 		try {
 			algorithm.postProcess();
